그룹 통계 입력값 int->double 수정

+++ LifeMaster-BE/src/main/java/com/example/LifeMaster_BE/Group/GoalProgress/GoalProgressResponseDTO.java
@@ -30,7 +30,7 @@ public class GoalProgressResponseDTO {
     private String goalName;
 
     @Schema(description = "진행 값", example = "2")
-    private int progressValue;
+    private double progressValue;
 
     @Schema(description = "제출 시간")
     private LocalDateTime submittedAt;
